+++ stdlib/IterativeEigenSolvers/test/runtests.jl
@@ -13,8 +13,8 @@ using Test
 
         testtol = 1e-6
 
-        @testset for elty in (Float64, Complex128)
-            if elty == Complex64 || elty == Complex128
+        @testset for elty in (Float64, ComplexF64)
+            if elty == ComplexF32 || elty == ComplexF64
                 a = acmplx
                 b = bcmplx
             else
@@ -90,7 +90,7 @@ using Test
                 if elty == Float64
                     @test_throws ArgumentError eigs(a+a.',which=:SI)
                     @test_throws ArgumentError eigs(a+a.',which=:LI)
-                    @test_throws ArgumentError eigs(a,sigma=rand(Complex64))
+                    @test_throws ArgumentError eigs(a,sigma=rand(ComplexF32))
                 end
             end
         end