+++ test/linalg/matmul.jl
@@ -16,8 +16,8 @@ using Base.LinAlg: mul!, Adjoint, Transpose
     @test ones(0,0)*ones(0,0) == zeros(0,0)
     @test Matrix{Float64}(uninitialized, 5, 0) |> t -> t't == zeros(0,0)
     @test Matrix{Float64}(uninitialized, 5, 0) |> t -> t*t' == zeros(5,5)
-    @test Matrix{Complex128}(uninitialized, 5, 0) |> t -> t't == zeros(0,0)
-    @test Matrix{Complex128}(uninitialized, 5, 0) |> t -> t*t' == zeros(5,5)
+    @test Matrix{ComplexF64}(uninitialized, 5, 0) |> t -> t't == zeros(0,0)
+    @test Matrix{ComplexF64}(uninitialized, 5, 0) |> t -> t*t' == zeros(5,5)
 end
 @testset "2x2 matmul" begin
     AA = [1 2; 3 4]
@@ -167,7 +167,7 @@ end
     @test mul!(sC, Transpose(A), B) == A'*B
 
     Aim = A .- im
-    C = zeros(Complex128,8,8)
+    C = zeros(ComplexF64,8,8)
     sC = view(C, 1:2:8, 1:2:8)
     B = reshape(map(Float64,-9:10),5,4) .+ im
     @test mul!(sC, Adjoint(Aim), Aim) == Aim'*Aim
@@ -215,7 +215,7 @@ end
 # issue #6450
 @test dot(Any[1.0,2.0], Any[3.5,4.5]) === 12.5
 
-@testset "dot" for elty in (Float32, Float64, Complex64, Complex128)
+@testset "dot" for elty in (Float32, Float64, ComplexF32, ComplexF64)
     x = convert(Vector{elty},[1.0, 2.0, 3.0])
     y = convert(Vector{elty},[3.5, 4.5, 5.5])
     @test_throws DimensionMismatch dot(x, 1:2, y, 1:3)
@@ -262,7 +262,7 @@ end
 
 @test_throws ArgumentError Base.LinAlg.copytri!(ones(10,10),'Z')
 
-@testset "gemv! and gemm_wrapper for $elty" for elty in [Float32,Float64,Complex128,Complex64]
+@testset "gemv! and gemm_wrapper for $elty" for elty in [Float32,Float64,ComplexF64,ComplexF32]
     @test_throws DimensionMismatch Base.LinAlg.gemv!(ones(elty,10),'N',rand(elty,10,10),ones(elty,11))
     @test_throws DimensionMismatch Base.LinAlg.gemv!(ones(elty,11),'N',rand(elty,10,10),ones(elty,10))
     @test Base.LinAlg.gemv!(ones(elty,0),'N',rand(elty,0,0),rand(elty,0)) == ones(elty,0)